Update book information display and add reading options
This update includes changes on how the book information is displayed on the BookPage.vue component. A book can now be directly read or downloaded in various formats offered, thanks to the addition of read and download buttons. Fetch processes for the book data was also streamlined. Changes to the Book.vue component were made to make the displayed book titles and subjects more concise. HeaderPage.vue contained minor alterations to optimize UI.